# eventify/ eventify-template/eventify-backend/src/main/java/com/eventify/eventify_backend/api/dto/response/ErrorResponse.java
package com.eventify.eventify_backend.api.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ErrorResponse {

    private String code;
    private String level;
    private String message;
    private List<ErrorDetail> details;


    @Data
    @Builder
    @NoArgsConstructor
    @AllArgsConstructor
    public static class ErrorDetail {
        private String field;
        private String message;
    }
}